Fix: refatorando funções com complexidades acima de 5
